+++ SimulationCore/src/com/aexiz/daviz/simulation/algorithm/AbstractBasicAlgorithm.java
@@ -1,4 +1,0 @@
-package com.aexiz.daviz.simulation.algorithm;
-
-public abstract class AbstractBasicAlgorithm extends AbstractAlgorithm implements BasicAlgorithm {
-}