+++ SimulationFrege/src/com/aexiz/daviz/simulation/algorithm/AbstractFregeControlAlgorithm.java
@@ -1,4 +1,4 @@
 package com.aexiz.daviz.simulation.algorithm;
 
-public abstract class AbstractFregeControlAlgorithm extends AbstractControlAlgorithm implements FregeAlgorithm {
+public abstract class AbstractFregeControlAlgorithm extends AbstractAlgorithm implements FregeAlgorithm, BasicAlgorithm {
 }